requirements.txt
datamapplot>=0.4.0
umap-learn
lancedb
pandas
numpy
requests
tqdm
pyarrow
